+++ packages/ui-assets/src/types.d.ts
@@ -1,8 +1,0 @@
-// Copyright 2017-2023 @polkadot/ui-assets authors & contributors
-// SPDX-License-Identifier: Apache-2.0
-
-declare module '*.svg' {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const content: any;
-  export default content;
-}